projects/camera-mcp/pyproject.toml
[project]
name = "distiller-cm5-cam-mcp"
version = "1.0.0"
description = "Pi Camera MCP Server - Provides camera snapshot and status tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.1.2",
    "picamera2>=0.3.12",
    "pillow>=9.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "distiller-mcp-hub",
]
[[project.authors]]
name = "Your Name"
email = "your.email@example.com"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
distiller-cm5-cam-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
distiller-mcp-hub = { path = "../../", editable = true }

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto" 
